\name{pixelWidth}
\alias{pixelWidth}
\title{
Pixel width of a habitat mask
}
\description{
A simple utility function to extract the pixel width attribute from an object of class \code{JAGSmask}.
}
\usage{
pixelWidth(JAGSmask)
}
\arguments{
  \item{JAGSmask}{
An object of class \code{JAGSmask}.
}
}
\value{
A scalar; the value of the pixelWidth attribute in metres.
}
\author{
Mike Meredith
}
\seealso{
\code{\link{convertMask}} 
}
